refactor(user permissions table): split out components to reduce file size
